Move shader hotswap & tile info to devtools class
